# apps/web/lib/api/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "AssignmentPayload": {
        "properties": {
          "assignment_id": {
            "pattern": "^asg_",
            "title": "Assignment Id",
            "type": "string"
          },
          "chunk_id": {
            "pattern": "^chk_",
            "title": "Chunk Id",
            "type": "string"
          },
          "deadline_at": {
            "format": "date-time",
            "title": "Deadline At",
            "type": "string"
          },
          "input_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Input Sha256",
            "type": "string"
          },
          "input_url": {
            "title": "Input Url",
            "type": "string"
          },
          "kernel": {
            "title": "Kernel",
            "type": "string"
          },
          "kernel_version": {
            "title": "Kernel Version",
            "type": "string"
          },
          "n_features": {
            "minimum": 1.0,
            "title": "N Features",
            "type": "integer"
          },
          "n_rows": {
            "minimum": 1.0,
            "title": "N Rows",
            "type": "integer"
          },
          "params": {
            "additionalProperties": true,
            "title": "Params",
            "type": "object"
          },
          "task_id": {
            "pattern": "^tsk_",
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "assignment_id",
          "task_id",
          "chunk_id",
          "kernel",
          "kernel_version",
          "params",
          "input_url",
          "input_sha256",
          "n_rows",
          "n_features",
          "deadline_at"
        ],
        "title": "AssignmentPayload",
        "type": "object"
      },
      "Benchmark": {
        "properties": {
          "bench_version": {
            "title": "Bench Version",
            "type": "string"
          },
          "score_cells_per_sec": {
            "exclusiveMinimum": 0.0,
            "title": "Score Cells Per Sec",
            "type": "number"
          }
        },
        "required": [
          "score_cells_per_sec",
          "bench_version"
        ],
        "title": "Benchmark",
        "type": "object"
      },
...
      "CancelDirective": {
        "properties": {
          "assignment_id": {
            "title": "Assignment Id",
            "type": "string"
          },
          "type": {
            "const": "cancel",
            "default": "cancel",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "assignment_id"
        ],
        "title": "CancelDirective",
        "type": "object"
      },
...
      "DeviceCapabilities": {
        "properties": {
          "battery": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Battery"
          },
          "logical_cores": {
            "anyOf": [
              {
                "maximum": 1024.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logical Cores"
          },
          "memory_gb_reported": {
            "anyOf": [
              {
                "maximum": 1024.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Gb Reported"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "network_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Network Type"
          },
          "platform_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform Version"
          },
          "storage_quota_mb": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Quota Mb"
          }
        },
        "title": "DeviceCapabilities",
        "type": "object"
      },
...
      "DeviceState": {
        "enum": [
          "initializing",
          "idle",
          "busy"
        ],
        "title": "DeviceState",
        "type": "string"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "default": [],
            "items": {},
            "title": "Details",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "HeartbeatRequest": {
        "properties": {
          "battery": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Battery"
          },
          "current_assignment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Assignment Id"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/DeviceState"
          }
        },
        "required": [
          "session_id",
          "state"
        ],
        "title": "HeartbeatRequest",
        "type": "object"
      },
      "HeartbeatResponse": {
        "properties": {
          "directives": {
            "default": [],
            "items": {
              "discriminator": {
                "mapping": {
                  "cancel": "#/components/schemas/CancelDirective",
                  "refresh_runtime": "#/components/schemas/RefreshRuntimeDirective",
                  "reregister": "#/components/schemas/ReregisterDirective",
                  "run": "#/components/schemas/RunDirective"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/RunDirective"
                },
                {
                  "$ref": "#/components/schemas/CancelDirective"
                },
                {
                  "$ref": "#/components/schemas/RefreshRuntimeDirective"
                },
                {
                  "$ref": "#/components/schemas/ReregisterDirective"
                }
              ]
            },
            "title": "Directives",
            "type": "array"
          },
          "next_heartbeat_ms": {
            "minimum": 500.0,
            "title": "Next Heartbeat Ms",
            "type": "integer"
          },
          "server_time": {
            "format": "date-time",
            "title": "Server Time",
            "type": "string"
          }
        },
        "required": [
          "server_time",
          "next_heartbeat_ms"
        ],
        "title": "HeartbeatResponse",
        "type": "object"
      },
...
      "RefreshRuntimeDirective": {
        "properties": {
          "kernel_version": {
            "title": "Kernel Version",
            "type": "string"
          },
          "type": {
            "const": "refresh_runtime",
            "default": "refresh_runtime",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "kernel_version"
        ],
        "title": "RefreshRuntimeDirective",
        "type": "object"
      },
      "ReregisterDirective": {
        "properties": {
          "type": {
            "const": "reregister",
            "default": "reregister",
            "title": "Type",
            "type": "string"
          }
        },
        "title": "ReregisterDirective",
        "type": "object"
      },
...
      "RunDirective": {
        "properties": {
          "assignment": {
            "$ref": "#/components/schemas/AssignmentPayload"
          },
          "type": {
            "const": "run",
            "default": "run",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "assignment"
        ],
        "title": "RunDirective",
        "type": "object"
      },
      "RuntimeInfo": {
        "properties": {
          "bundle": {
            "title": "Bundle",
            "type": "string"
          },
          "kind": {
            "enum": [
              "pyodide",
              "cpython"
            ],
            "title": "Kind",
            "type": "string"
          },
          "numpy": {
            "title": "Numpy",
            "type": "string"
          },
          "pyodide": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pyodide"
          },
          "python": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Python"
          }
        },
        "required": [
          "kind",
          "numpy",
          "bundle"
        ],
        "title": "RuntimeInfo",
        "type": "object"
      },
      "RuntimeManifest": {
        "properties": {
          "heartbeat_busy_ms": {
            "title": "Heartbeat Busy Ms",
            "type": "integer"
          },
          "heartbeat_idle_ms": {
            "title": "Heartbeat Idle Ms",
            "type": "integer"
          },
          "kernel_bundle_sha256": {
            "title": "Kernel Bundle Sha256",
            "type": "string"
          },
          "kernel_bundle_version": {
            "title": "Kernel Bundle Version",
            "type": "string"
          },
          "numpy_version": {
            "title": "Numpy Version",
            "type": "string"
          },
          "offline_after_seconds": {
            "title": "Offline After Seconds",
            "type": "integer"
          },
          "pyodide_version": {
            "title": "Pyodide Version",
            "type": "string"
          }
        },
        "required": [
          "pyodide_version",
          "numpy_version",
          "kernel_bundle_version",
          "kernel_bundle_sha256",
          "heartbeat_idle_ms",
          "heartbeat_busy_ms",
          "offline_after_seconds"
        ],
        "title": "RuntimeManifest",
        "type": "object"
      },
      "SessionRequest": {
        "properties": {
          "benchmark": {
            "$ref": "#/components/schemas/Benchmark"
          },
          "capabilities": {
            "$ref": "#/components/schemas/DeviceCapabilities",
            "default": {}
          },
          "runtime": {
            "$ref": "#/components/schemas/RuntimeInfo"
          }
        },
        "required": [
          "runtime",
          "benchmark"
        ],
        "title": "SessionRequest",
        "type": "object"
      },
      "SessionResponse": {
        "properties": {
          "heartbeat_busy_ms": {
            "title": "Heartbeat Busy Ms",
            "type": "integer"
          },
          "heartbeat_idle_ms": {
            "title": "Heartbeat Idle Ms",
            "type": "integer"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "heartbeat_idle_ms",
          "heartbeat_busy_ms"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ProofNet API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/runtime/kernels/{version}": {
      "get": {
        "operationId": "kernel_bundle_api_v1_runtime_kernels__version__get",
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Kernel Bundle",
        "tags": [
          "worker"
        ]
      }
    },
    "/api/v1/runtime/manifest": {
      "get": {
        "operationId": "runtime_manifest_api_v1_runtime_manifest_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Runtime Manifest",
        "tags": [
          "worker"
        ]
      }
    },
...
    "/api/v1/worker/heartbeat": {
      "post": {
        "operationId": "heartbeat_api_v1_worker_heartbeat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Heartbeat",
        "tags": [
          "worker"
        ]
      }
    },
    "/api/v1/worker/session": {
      "post": {
        "operationId": "start_session_api_v1_worker_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Start Session",
        "tags": [
          "worker"
        ]
      }
    }
  }
}